+++ docs/conf.py
@@ -58,7 +58,7 @@ master_doc = 'index'
 # General information about the project.
 project = u'MLConjug3'
 author = u'SekouDiaoNlp'
-copyright = u"2021, SekouDiaoNlp"
+copyright = u"2023, SekouDiaoNlp"
 
 # The version info for the project you're documenting, acts as replacement
 # for |version| and |release|, also used in various other places throughout
@@ -117,8 +117,7 @@ gettext_compact = False
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-html_theme = "sphinx_rtd_theme"
-
+html_theme = "furo"
 # Theme options are theme-specific and customize the look and feel of a
 # theme further.  For a list of options available for each theme, see the
 # documentation.