+++ mlconjug3/PyVerbiste/__init__.py
@@ -1,0 +1,3 @@
+from .PyVerbiste import Verbiste
+
+__all__ = ['Verbiste', ]