+++ mlconjug3/__init__.py
@@ -26,13 +26,14 @@ Usage example:
 
 __author__ = """SekouDiaoNlp"""
 __email__ = 'diao.sekou.nlp@gmail.com'
-__version__ = '3.8.3'
+__version__ = '3.9.0'
 __copyright__ = "Copyright (c) 2021, SekouDiaoNlp"
 __credits__ = ("Sekou Diao", "Pierre Sarrazin",)
 __license__ = "MIT"
 __maintainer__ = "SekouDiaoNlp"
 __status__ = "Production"
 
+from .constants import *
 from .mlconjug import *
 from .PyVerbiste import *
 
@@ -47,13 +48,6 @@ import platform
 from locale import windows_locale, getdefaultlocale
 import gettext
 import inspect
-
-# Sets up the automatic translation of annotated strings displayed to the user.
-_RESOURCE_PACKAGE = 'mlconjug3'
-_TRANSLATIONS_PATH = pkg_resources.resource_filename(_RESOURCE_PACKAGE, 'locale')
-
-_SUPPORTED_LANGUAGES = ('default', 'en', 'es', 'fr', 'it', 'pt', 'ro')
-_TRANSLATED_LANGUAGES = _SUPPORTED_LANGUAGES[2:]
 
 
 def _get_user_locale():
@@ -100,14 +94,14 @@ def _getdoc(obj):
 
 _user_locale = _get_user_locale()
 
-if _user_locale in _TRANSLATED_LANGUAGES:
-    _MLCONJUG_TRANSLATIONS = gettext.translation(domain='mlconjug3',
-                                                 localedir=_TRANSLATIONS_PATH,
+if _user_locale in TRANSLATED_LANGUAGES:
+    MLCONJUG_TRANSLATIONS = gettext.translation(domain='mlconjug3',
+                                                 localedir=TRANSLATIONS_PATH,
                                                  languages=[_user_locale], fallback=True, codeset='UTF-8')
 else:
-    _MLCONJUG_TRANSLATIONS = gettext.NullTranslations()
+    MLCONJUG_TRANSLATIONS = gettext.NullTranslations()
 
-_MLCONJUG_TRANSLATIONS.install()
+MLCONJUG_TRANSLATIONS.install()
 
 # Replaces the getdoc method
 inspect.getdoc = _getdoc