+++ mlconjug3/conjug_manager/__init__.py
@@ -1,0 +1,3 @@
+from .conjug_manager import ConjugManager
+
+__all__ = ['ConjugManager',]