+++ mlconjug3/dataset/__init__.py
@@ -1,0 +1,2 @@
+from .dataset import DataSet
+__all__ = ['DataSet', ]