+++ mlconjug3/feature_extractor/__init__.py
@@ -1,0 +1,5 @@
+from .feature_extractor import extract_verb_features
+
+from mlconjug3.constants import *
+
+__all__ = ['extract_verb_features', ]