+++ mlconjug3/models/__init__.py
@@ -1,0 +1,3 @@
+from .models import Model
+
+__all__ = ['Model', ]