+++ mlconjug3/utils/__init__.py
@@ -1,0 +1,4 @@
+from .logger import logger
+from .model_trainer import ConjugatorTrainer
+
+__all__ = ['logger', 'ConjugatorTrainer',]