+++ mlconjug3/verbs/__init__.py
@@ -1,0 +1,3 @@
+from .verbs import VerbInfo, Verb, VerbFr, VerbEn, VerbEs, VerbIt, VerbPt, VerbRo
+
+__all__ = ['VerbInfo', 'Verb', 'VerbFr', 'VerbEn', 'VerbEs', 'VerbIt', 'VerbPt', 'VerbRo',]